move definition of contentScripts to the top
Decided to move the scripts array to the top background.js and rename it
to contentScripts in order to be more clear and to possible allow for
dynamic adding of content scripts. For example a setting in the options
page could trigger background.js to push another script into the array
extending functionality but not forcing the script to be always loaded
unless indicated by the user.